+++ lib/factual/api.rb
@@ -3,7 +3,7 @@ require 'cgi'
 
 class Factual
   class API
-    VERSION = "1.3.1"
+    VERSION = "1.3.2"
     API_V3_HOST = "api.v3.factual.com"
     DRIVER_VERSION_TAG = "factual-ruby-driver-v" + VERSION
     PARAM_ALIASES = { :search => :q, :sort_asc => :sort }