+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * added integration test for unicode queries
 * added insert
 * changed factual-auth yaml file from under the integration test folder to ~/.factual
+* added clear
+* timeout
 
 ## v1.3.1
 * removed sugar crosswalk, switch to normal table