+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * added insert
 * changed factual-auth yaml file from under the integration test folder to ~/.factual
 * added clear
-* timeout
+* timeout support
 
 ## v1.3.1
 * removed sugar crosswalk, switch to normal table